refactor: Refactor app directory structure to be able to use bloc
