+++ src/support_sphere/lib/constants/color.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+/// Color constants used in the app
+abstract class ColorConstants {
+  static const Color seed = Color.fromARGB(255, 14, 54, 70);
+  static const Color transparent = Color.fromARGB(0, 0, 0, 0);
+}
