+++ src/support_sphere/lib/constants/environment.dart
@@ -1,0 +1,12 @@
+/// Environment variables constants.
+abstract class EnvironmentConstants {
+  static const String supabaseUrl = String.fromEnvironment(
+    'SUPABASE_URL',
+    defaultValue: 'http://localhost',
+  );
+
+  static const String supabaseAnnonKey = String.fromEnvironment(
+    'SUPABASE_ANNON_KEY',
+    defaultValue: 'your-super-secret-jwt-token-with-at-least-32-characters-long',
+  );
+}
