+++ src/support_sphere/lib/presentation/components/auth/text_form_builder.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:support_sphere/components/custom_card.dart';
+import 'package:support_sphere/presentation/components/custom_card.dart';
 
 class TextFormBuilder extends StatefulWidget {
   final String? initialValue;
