+++ src/support_sphere/lib/presentation/components/custom_card.dart
