+++ src/support_sphere/lib/utils/supabase.dart
@@ -1,0 +1,3 @@
+import 'package:supabase_flutter/supabase_flutter.dart';
+
+final supabaseClient = Supabase.instance.client;
