+++ src/support_sphere/pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   loading_overlay: ^0.3.0
   provider: ^6.1.2
   google_fonts: ^6.2.1
+  supabase_flutter: ^2.6.0
 
 dev_dependencies:
   flutter_test:
